separate page server and client components
